+++ src/components/product/product.ts
@@ -3,8 +3,8 @@ import Button from '../button/button';
 import { Price, ProductData, TypedMoney } from '@commercetools/platform-sdk';
 import formatPrice from '../../utils/formatPrice';
 import Notification from '../notification/notification';
-import addToCart from '../../services/addToCart';
-import removeProductFromCart from '../../services/removeProductFromCart';
+import addToCart from '../../services/basket/addToCart';
+import removeProductFromCart from '../../services/basket/removeProductFromCart';
 import cartStore from '../../store/cart-store';
 import './product.scss';
