+++ src/services/basket/createUserCart.ts
@@ -1,5 +1,5 @@
 import { Cart, ClientResponse } from '@commercetools/platform-sdk';
-import apiExistingToken from './apiExistingToken';
+import apiExistingToken from '../apiExistingToken';
 
 export default async function createUserCart(token: string): Promise<Cart> {
   const response: ClientResponse<Cart> = await apiExistingToken(token)
