+++ src/services/getDiscountCode.ts
@@ -1,10 +1,10 @@
 import { Cart, ClientResponse, DiscountCode } from '@commercetools/platform-sdk';
 import userStore from '../store/user-store';
-import createAnonymusCart from './createAnonymusCart';
-import getActiveCart from './getActiveCart';
+import createAnonymusCart from './basket/createAnonymusCart';
+import getActiveCart from './basket/getActiveCart';
 import apiExistingToken from './apiExistingToken';
 import Notification from '../components/notification/notification';
-import createUserCart from './createUserCart';
+import createUserCart from './basket/createUserCart';
 
 export default async function getDiscountCode(): Promise<DiscountCode | undefined> {
   const { isAuth } = userStore.getState();
